# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Catálogo de Ferias - API canónica
  version: 1.0.0
  description: Contrato HTTP estable del backend de Unidades Productivas y ferias.
servers:
  - url: http://localhost:5000/api
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      security: []
      tags: [Auth]
      summary: Iniciar sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, password]
              properties:
                login: {type: string}
                password: {type: string, format: password}
      responses:
        '200':
          description: Tokens emitidos
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TokenPair'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Renovar token de acceso
      responses:
        '200': {description: Token de acceso renovado}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /auth/me:
    get:
      tags: [Auth]
      summary: Consultar sesión
      responses:
        '200': {description: Usuario autenticado}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /auth/reauthenticate:
    post:
      tags: [Auth]
      summary: Confirmar contraseña actual
      requestBody: {$ref: '#/components/requestBodies/JsonBody'}
      responses:
        '200': {description: Sesión confirmada}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /auth/change-password:
    post:
      tags: [Auth]
      summary: Cambiar contraseña
      requestBody: {$ref: '#/components/requestBodies/JsonBody'}
      responses:
        '200': {description: Contraseña actualizada y sesiones anteriores revocadas}
        '400': {$ref: '#/components/responses/BadRequest'}
...
  /auth/forgot-password:
    post:
      security: []
      tags: [Auth]
      summary: Solicitar código de recuperación
      requestBody: {$ref: '#/components/requestBodies/JsonBody'}
      responses:
        '200': {description: Respuesta uniforme}
  /auth/verify-recovery-code:
    post:
      security: []
      tags: [Auth]
      summary: Verificar código de recuperación
      requestBody: {$ref: '#/components/requestBodies/JsonBody'}
      responses:
        '200': {description: Token de restablecimiento emitido}
        '400': {$ref: '#/components/responses/BadRequest'}
  /auth/reset-password:
    post:
      security: []
      tags: [Auth]
      summary: Restablecer contraseña
      requestBody: {$ref: '#/components/requestBodies/JsonBody'}
      responses:
        '200': {description: Contraseña restablecida}
        '400': {$ref: '#/components/responses/BadRequest'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  requestBodies:
    JsonBody:
      required: true
      content:
        application/json:
          schema: {type: object, additionalProperties: true}
...
  responses:
    BadRequest:
      description: Datos inválidos
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Unauthorized: {description: Token o credenciales inválidos}
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: {type: string}
        details: {type: object, additionalProperties: true}
    TokenPair:
      type: object
      required: [access_token, refresh_token, user]
      properties:
        access_token: {type: string}
        refresh_token: {type: string}
        user: {type: object, additionalProperties: true}
...